lightningtw-cat/requirements.txt
# For Telegram Bot
python-telegram-bot

# For Google Sheets API
gspread
oauth2client

# For Data Analysis and Reporting
pandas
matplotlib
seaborn

# For loading .env files
python-dotenv

# For AI Model Training with PyTorch Lightning
torch
pytorch-lightning

# For Webhook Server (optional)
Flask

# For API Server
fastapi
uvicorn[standard]
pydantic
